requirements.txt
django
psycopg2-binary
django-tempus-dominus
django-recaptcha
whitenoise
pillow
passporteye
redis
celery

freezegun

gunicorn
django-sendgrid-v5
sentry-sdk
newrelic

coverage
flake8
pylint
pylint-django
django-debug-toolbar
selenium
